refactor: 우승자 찾는 기능를 CarFleet에 위임으로 변경
