+++ src/main/kotlin/game/domain/History.kt
@@ -8,4 +8,10 @@ class History(
         newRounds.add(cars)
         return History(newRounds)
     }
+
+    fun findWinners(): List<String> {
+        val lastRound = rounds.last()
+        val maxPosition = lastRound.maxOf { it.position }
+        return lastRound.filter { it.isSamePosition(maxPosition) }.map { it.name }
+    }
 }
